backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.6
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
Django==3.1.7
django-autoslug==1.9.8
django-bulk-sync==3.2.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-phonenumber-field==5.0.0
djangorestframework==3.12.4
Faker==8.1.0
hiredis==1.1.0
http-ece==1.1.0
hyperlink==21.0.0
idna
incremental==21.3.0
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
phonenumberslite==8.12.20
Pillow==8.1.2
psycopg2==2.8.6
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.3.17
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zope.interface==5.3.0
